+++ wodoo/click_config.py
@@ -1,6 +1,8 @@
 import os
+from collections import ChainMap
 import sys
 from pathlib import Path
+import importlib
 
 
 class Config(object):
@@ -16,24 +18,32 @@ class Config(object):
         def __exit__(self, type, value, traceback):
             self.config.force = self.force
 
-    def __init__(self, quiet=False, project_name=None, force=False, verbose=False, version=None):
+    def __init__(
+        self, quiet=False, project_name=None, force=False, verbose=False, version=None
+    ):
         from .consts import YAML_VERSION
         from . import odoo_config  # NOQA
 
-        from .init_functions import _get_customs_root
+        from .tools import _get_customs_root
 
         self._WORKING_DIR = _get_customs_root(Path(os.getcwd()))
         self._host_run_dir = None
-        self.project_name = project_name
+        self._project_name = None
         self.YAML_VERSION = YAML_VERSION
-        self._verbose = verbose
+        self.verbose = verbose
         self.force = force
         self.compose_version = YAML_VERSION
         self.quiet = quiet
         self.restrict = {}
+        self.dirs = {}
+        self.files = {}
+        self.commands = {}
+
+        self.project_name = project_name
 
     def _collect_files(self, files):
         import click
+
         for test in files:
             test = Path(test)
             if not test.exists():
@@ -52,7 +62,7 @@ class Config(object):
     @verbose.setter
     def verbose(self, value):
         self._verbose = value
-        os.environ['WODOO_VERBOSE'] = "1"
+        os.environ["WODOO_VERBOSE"] = "1" if value else "0"
 
     @property
     def WORKING_DIR(self):
@@ -61,8 +71,7 @@ class Config(object):
     @WORKING_DIR.setter
     def WORKING_DIR(self, value):
         self._WORKING_DIR = value
-        os.environ['CUSTOMS_DIR'] = self._WORKING_DIR
-
+        os.environ["CUSTOMS_DIR"] = self._WORKING_DIR
 
     @property
     def project_name(self):
@@ -72,11 +81,16 @@ class Config(object):
     def project_name(self, value):
         self._project_name = value
         if self._project_name:
-            os.environ["PROJECT_NAME"] = value
-            self.HOST_RUN_DIR = Path(os.environ["HOME"]) / ".odoo" / "run" / value
-        else:
-            os.environ["PROJECT_NAME"] = ""
-        self.setup_files_and_folders()
+            self.HOST_RUN_DIR = (
+                Path(os.environ["HOME"]) / ".odoo" / "run" / self._project_name
+            )
+        os.environ["PROJECT_NAME"] = self._project_name or ""
+        os.environ["project_name"] = self._project_name or ""
+        self._setup_files_and_folders()
+        os.environ["docker_compose"] = str(self.files.get("docker_compose")) or ""
+        self.load_dynamic_modules()
+        if self.verbose:
+            print(self.files["docker_compose"])
 
     @property
     def HOST_RUN_DIR(self):
@@ -85,22 +99,8 @@ class Config(object):
     @HOST_RUN_DIR.setter
     def HOST_RUN_DIR(self, value):
         self._host_run_dir = value
-        if value:
-            os.environ["HOST_RUN_DIR"] = str(value)
-        self.setup_files_and_folders()
-
-    def setup_files_and_folders(self):
-        from .init_functions import get_use_docker
-        from . import odoo_config  # NOQA
-
-        self.dirs = {}
-        self.files = {}
-        self.commands = {}
-
-        self.use_docker = get_use_docker(self.files)
-        from .init_functions import make_absolute_paths
-
-        make_absolute_paths(self, self.dirs, self.files, self.commands)
+        os.environ["HOST_RUN_DIR"] = str(value)
+        self._setup_files_and_folders()
 
     def forced(self):
         return Config.Forced(self)
@@ -153,3 +153,89 @@ class Config(object):
         )
         conn = DBConnection(self.dbname, host, port, user, password)
         return conn
+
+    @property
+    def use_docker(self):
+        from .myconfigparser import MyConfigParser  # NOQA
+
+        try:
+            myconfig = MyConfigParser(self.files["settings"])
+        except Exception:
+            USE_DOCKER = True
+        else:
+            USE_DOCKER = myconfig.get("USE_DOCKER", "1") == "1"
+
+        return USE_DOCKER
+
+    def _setup_files_and_folders(self):
+        from . import odoo_config  # NOQA
+
+        from .consts import default_dirs, default_files, default_commands
+
+        for (input, output) in [
+            (default_dirs, self.dirs),
+            (default_files, self.files),
+            (default_commands, self.commands),
+        ]:
+            output.clear()
+            for k, v in input.items():
+                output[k] = v
+
+        self.dirs["odoo_home"] = Path(os.environ["ODOO_HOME"])
+
+        def replace_keys(value, key_values):
+            for k, v in key_values.items():
+                p = ["${" + k + "}", f"${k}"]
+                for p in p:
+                    if p in str(value):
+                        value = value.replace(p, str(v))
+            return value
+
+        def make_absolute(d, key_values={}):
+            for k, v in list(d.items()):
+                if not v:
+                    continue
+                skip = False
+                v = replace_keys(v, key_values)
+
+                for value, name in [
+                    (self.HOST_RUN_DIR, "${run}"),
+                    (self.WORKING_DIR, "${working_dir}"),
+                    (self.project_name, "${project_name}"),
+                ]:
+                    if name in str(v):
+                        if value:
+                            v = str(v).replace(name, str(value))
+                        else:
+                            del d[k]
+                            skip = True
+                            break
+                if skip:
+                    continue
+                if str(v).startswith("~"):
+                    v = Path(os.path.expanduser(str(v)))
+
+                if not str(v).startswith("/"):
+                    v = self.dirs["odoo_home"] / v
+                d[k] = Path(v)
+
+        make_absolute(self.dirs)
+        make_absolute(self.files, self.dirs)
+
+        for k in self.commands:
+            self.commands[k] = [
+                replace_keys(x, ChainMap(self.__dict__, self.files, self.dirs))
+                for x in self.commands[k]
+            ]
+
+    def load_dynamic_modules(self):
+        parent_dir = self.dirs["images"]
+        for module in parent_dir.glob("*/__commands.py"):
+            if module.is_dir():
+                continue
+            spec = importlib.util.spec_from_file_location(
+                "dynamic_loaded_module",
+                str(module),
+            )
+            module = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(module)
