+++ wodoo/consts.py
@@ -1,5 +1,6 @@
 from pathlib import Path
-from .init_functions import _search_path
+from .tools import _search_path
+
 VERSIONS = [7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0]
 YAML_VERSION = '3.7'
 DEFAULT_IMAGES_REPO = "https://github.com/marcwimmer/wodoo-images"
