+++ wodoo/lib_control_with_docker.py
@@ -242,7 +242,6 @@ def build(ctx, config, machines=[], pull=False, no_cache=False, push=False):
     """
     no parameter all machines, first parameter machine name and passes other params; e.g. ./odoo build asterisk --no-cache"
     """
-    import pudb;pudb.set_trace()
     options = []
     if pull:
         options += ["--pull"]
